standardize all console scripts on `uoft-` namespace, plus bugfixes
